requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
gunicorn==21.2.0
reportlab==4.0.7
werkzeug==3.0.1
Flask-Login
Flask-SQLAlchemy
opencv-python
reportlab
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
reportlab==4.0.7
werkzeug==3.0.1
